extract svg to file
