+++ src/components/Header.jsx
@@ -1,11 +1,11 @@
 'use client'
 
-import { Popover } from '@headlessui/react'
-import { AnimatePresence, motion } from 'framer-motion'
 import { Button } from '@/components/base/Button'
 import { Container } from '@/components/base/Container'
-import { AppLogo } from '@/components/Logos'
 import { NavLinks } from '@/components/NavLinks'
+import AppLogo from '@/images/logos/app.svg'
+import { Popover } from '@headlessui/react'
+import { AnimatePresence, motion } from 'framer-motion'
 
 function MenuIcon(props) {
   return (
