--- src/components/Logos.jsx
@@ -1,162 +1,0 @@
-export function AppLogomark(props) {
-  return (
-    <svg viewBox="0 0 400 400" aria-hidden="true" {...props}>
-      <g fill={props.fill || '#000000'} fillOpacity="0.5">
-        <path d="M60.19,129.786 C62.223,154.36 83.178,172.297 107.319,172.909 C113.832,172.909 120.225,171.581 126.155,169.006 L171.847,215.667 C171.407,216.449 170.986,217.24 170.59,218.043 L110.793,209.985 C106.676,187.856 87.243,171.349 64.228,171.349 C38.07,171.349 16.856,192.566 16.855,218.722 C16.856,244.876 38.07,266.095 64.228,266.095 C81.814,266.095 97.77,256.375 105.95,241.168 L165.758,249.223 C166.304,252.812 167.221,256.335 168.497,259.732 L143.408,281.822 C140.784,281.056 138.057,280.665 135.304,280.665 C119.343,280.665 106.397,293.613 106.397,309.574 C106.397,325.533 119.343,338.479 135.304,338.479 C151.26,338.479 164.206,325.533 164.206,309.574 C164.206,308.313 164.124,307.056 163.96,305.81 L188.044,284.601 C196.711,290.528 206.953,293.707 217.565,293.707 C219.651,293.707 221.739,293.583 223.81,293.335 L245.079,335.789 C236.908,344.521 232.308,356.027 232.308,368.142 C232.301,380.604 236.829,390.557 244.627,400 L25.874,400 C13.634,400 3.711,390.077 3.711,377.837 L3.711,194.92 C3.711,189.001 6.078,183.328 10.285,179.165 L60.19,129.786 z"/>
-      </g>
-      <g fill={props.fill || '#ea9122'} fillOpacity="0.5">
-        <path d="M308.102,98.412 L389.715,179.165 C393.922,183.328 396.289,189.001 396.289,194.92 L396.289,377.837 C396.289,390.077 386.366,400 374.126,400 L313.703,400 C321.716,391.135 325.721,380.003 326.021,368.142 C326.022,341.983 304.805,320.767 278.649,320.767 C276.425,320.767 274.206,320.921 272.007,321.233 L251.512,280.319 C262.538,270.437 268.938,256.323 268.938,241.296 C268.938,223.395 259.839,206.933 245.079,197.342 L263.797,135.661 C284.845,134.686 303.93,119.34 308.102,98.412 z"/>
-        <path d="M222.558,113.178 C225.728,118.311 229.853,122.802 234.701,126.395 L215.725,188.919 C208.458,189.172 201.365,190.918 194.847,194.057 L149.28,147.53 C152.089,142.185 153.838,136.37 154.449,130.356 L222.558,113.178 z"/>
-        <path d="M200.516,0 L202.629,0.05 C207.875,0.561 212.242,2.9 216.108,6.359 L252.288,42.159 C232.964,46.857 219.012,62.448 215.94,82 L146.852,99.424 C138.888,87.343 126.215,80.259 112.047,78.475 L184.93,6.359 C188.604,2.811 193.3,0.539 198.408,0.05 L200.516,0 z"/>
-      </g>
-      <g fill={props.fill || '#000000'}>
-        <path d="M60.19,129.786 C62.223,154.36 83.178,172.297 107.319,172.909 C113.832,172.909 120.225,171.581 126.155,169.006 L171.847,215.667 C171.407,216.449 170.986,217.24 170.59,218.043 L110.793,209.985 C106.676,187.856 87.243,171.349 64.228,171.349 C38.07,171.349 16.856,192.566 16.855,218.722 C16.856,244.876 38.07,266.095 64.228,266.095 C81.814,266.095 97.77,256.375 105.95,241.168 L165.758,249.223 C166.304,252.812 167.221,256.335 168.497,259.732 L143.408,281.822 C140.784,281.056 138.057,280.665 135.304,280.665 C119.343,280.665 106.397,293.613 106.397,309.574 C106.397,325.533 119.343,338.479 135.304,338.479 C151.26,338.479 164.206,325.533 164.206,309.574 C164.206,308.313 164.124,307.056 163.96,305.81 L188.044,284.601 C196.711,290.528 206.953,293.707 217.565,293.707 C219.651,293.707 221.739,293.583 223.81,293.335 L245.078,335.789 C236.908,344.521 232.308,356.027 232.308,368.142 C232.301,380.604 236.829,390.557 244.627,400 L25.874,400 C13.634,400 3.711,390.077 3.711,377.837 L3.711,194.92 C3.711,189.001 6.078,183.328 10.285,179.165 L60.19,129.786 z M114.505,260.482 C105.513,271.915 91.664,279.806 77.522,282.731 C71.37,284.003 69.59,283.911 64.228,284.095 C48.444,284.18 33.77,278.246 21.711,268.294 L21.711,377.837 C21.711,380.136 23.575,382 25.874,382 L215.797,382 C214.685,377.458 214.407,372.796 214.308,368.142 C214.383,356.119 217.587,344.152 224.03,333.96 L212.777,311.498 C205.109,311.183 197.717,309.107 190.619,306.318 L181.945,313.956 C179.901,338.213 159.105,355.877 135.304,356.479 C109.402,356.479 88.397,335.475 88.397,309.574 C88.727,284.389 107.153,265.336 131.905,262.826 L114.505,260.482 z M124.247,192.785 C124.429,193.162 124.293,192.871 124.6,193.683 L125.207,193.765 L124.247,192.785 z"/>
-      </g>
-      <g fill={props.fill || '#ea9122'}>
-        <path d="M308.102,98.412 L389.714,179.165 C393.922,183.328 396.289,189.001 396.289,194.92 L396.289,377.837 C396.289,390.077 386.366,400 374.126,400 L313.703,400 C321.716,391.135 325.721,380.003 326.021,368.142 C326.022,341.983 304.805,320.767 278.649,320.767 C276.425,320.767 274.206,320.921 272.007,321.233 L251.512,280.319 C262.538,270.437 268.938,256.323 268.938,241.296 C268.938,223.395 259.839,206.933 245.079,197.342 L263.797,135.661 C284.845,134.686 303.93,119.34 308.102,98.412 z M313.188,128.766 C304.115,139.605 291.792,148.591 277.753,151.66 L265.797,191.055 C279.667,204.909 286.324,221.841 286.938,241.296 C286.968,256.463 281.936,270.931 273.051,283.128 L282.981,302.95 C317.99,305.901 343.137,333.165 344.021,368.142 C344.021,368.294 344.02,368.446 344.016,368.598 C344.009,373.126 343.239,377.575 342.371,382 L374.126,382 C376.425,382 378.289,380.136 378.289,377.837 L378.289,194.92 C378.289,193.808 377.845,192.742 377.054,191.96 L313.188,128.766 z"/>
-        <path d="M222.558,113.177 C225.728,118.311 229.853,122.802 234.701,126.395 L215.725,188.919 C208.458,189.172 201.365,190.918 194.847,194.057 L149.279,147.529 C152.089,142.185 153.838,136.37 154.449,130.356 L222.558,113.177 z M213.58,134.005 L171.599,144.594 L199.687,173.274 L201.821,172.748 L213.58,134.005 z"/>
-        <path d="M200.516,-0 L202.629,0.05 C207.875,0.561 212.242,2.9 216.108,6.359 L252.288,42.159 C232.964,46.857 219.012,62.448 215.94,82 L146.852,99.424 C138.888,87.343 126.215,80.259 112.047,78.475 L184.93,6.359 C188.604,2.811 193.3,0.539 198.408,0.05 L200.516,-0 z M200.517,18.005 L199.847,18.021 C198.902,18.113 198.144,18.714 197.446,19.296 L144.398,71.787 C150.858,76.679 148.575,74.688 153.321,79.229 L201.185,67.158 C205.292,55.517 212.471,45.333 222.028,37.539 L203.91,19.613 C203.117,18.9 202.216,18.238 201.153,18.02 L200.517,18.005 z"/>
-      </g>
-    </svg>
-  )
-}
-
-export function AppLogo(props) {
-  return (
-    <svg viewBox="0 0 1216, 400" aria-hidden="true" {...props}>
-      <g>
-        <g fill={props.fill || '#000000'}  fillOpacity="0.5">
-          <path d="M56.479,129.786 C58.512,154.36 79.467,172.297 103.609,172.909 C110.121,172.909 116.515,171.581 122.444,169.006 L168.136,215.667 C167.696,216.449 167.275,217.24 166.879,218.043 L107.083,209.985 C102.965,187.856 83.532,171.349 60.518,171.349 C34.359,171.349 13.145,192.566 13.145,218.722 C13.145,244.876 34.359,266.095 60.518,266.095 C78.103,266.095 94.06,256.375 102.239,241.168 L162.047,249.223 C162.593,252.812 163.51,256.335 164.786,259.732 L139.698,281.822 C137.073,281.056 134.346,280.665 131.593,280.665 C115.632,280.665 102.686,293.613 102.686,309.574 C102.686,325.533 115.632,338.479 131.593,338.479 C147.549,338.479 160.5,325.533 160.5,309.574 C160.5,308.313 160.413,307.056 160.249,305.81 L184.334,284.601 C193,290.528 203.242,293.707 213.854,293.707 C215.94,293.707 218.028,293.583 220.1,293.335 L241.368,335.789 C233.198,344.521 228.598,356.027 228.598,368.142 C228.59,380.604 233.118,390.557 240.917,400 L22.164,400 C9.923,400 0,390.077 0,377.837 L0,194.92 C0,189.001 2.368,183.328 6.575,179.165 L56.479,129.786 z"/>
-        </g>
-        <g fill={props.fill || '#ea9122'}  fillOpacity="0.5">
-          <path d="M304.391,98.412 L386.004,179.165 C390.211,183.328 392.579,189.001 392.579,194.92 L392.579,377.837 C392.579,390.077 382.656,400 370.415,400 L309.992,400 C318.006,391.135 322.01,380.003 322.311,368.142 C322.311,341.983 301.094,320.767 274.938,320.767 C272.714,320.767 270.5,320.921 268.297,321.233 L247.801,280.319 C258.827,270.437 265.228,256.323 265.228,241.296 C265.228,223.395 256.128,206.933 241.368,197.342 L260.087,135.661 C281.135,134.686 300.219,119.34 304.391,98.412 z"/>
-          <path d="M218.847,113.178 C222.017,118.311 226.142,122.802 230.99,126.395 L212.014,188.919 C204.747,189.172 197.654,190.918 191.137,194.057 L145.569,147.53 C148.378,142.185 150.127,136.37 150.738,130.356 L218.847,113.178 z"/>
-          <path d="M196.805,0 L198.918,0.05 C204.164,0.561 208.531,2.9 212.397,6.359 L248.577,42.159 C229.253,46.857 215.302,62.448 212.229,82 L143.141,99.424 C135.178,87.343 122.504,80.259 108.336,78.475 L181.219,6.359 C184.893,2.811 189.589,0.539 194.697,0.05 L196.805,0 z"/>
-        </g>
-      </g>
-      <g fill={props.fill || '#000000'}>
-        <path d="M56.479,129.786 C58.512,154.36 79.467,172.297 103.609,172.909 C110.121,172.909 116.515,171.581 122.444,169.006 L168.136,215.667 C167.696,216.449 167.275,217.24 166.879,218.043 L107.083,209.985 C102.965,187.856 83.532,171.349 60.518,171.349 C34.359,171.349 13.145,192.566 13.145,218.722 C13.145,244.876 34.359,266.095 60.518,266.095 C78.103,266.095 94.06,256.375 102.239,241.168 L162.047,249.223 C162.593,252.812 163.51,256.335 164.786,259.732 L139.698,281.822 C137.073,281.056 134.346,280.665 131.593,280.665 C115.632,280.665 102.686,293.613 102.686,309.574 C102.686,325.533 115.632,338.479 131.593,338.479 C147.549,338.479 160.5,325.533 160.5,309.574 C160.5,308.313 160.413,307.056 160.249,305.81 L184.334,284.601 C193,290.528 203.242,293.707 213.854,293.707 C215.94,293.707 218.028,293.583 220.1,293.335 L241.368,335.789 C233.198,344.521 228.598,356.027 228.598,368.142 C228.59,380.604 233.118,390.557 240.917,400 L22.164,400 C9.923,400 0,390.077 0,377.837 L0,194.92 C0,189.001 2.368,183.328 6.575,179.165 L56.479,129.786 z M110.794,260.482 C101.802,271.915 87.953,279.806 73.811,282.731 C67.66,284.003 65.879,283.911 60.518,284.095 C44.733,284.18 30.059,278.246 18,268.294 L18,377.837 C18,380.136 19.864,382 22.164,382 L212.086,382 C210.975,377.458 210.696,372.796 210.598,368.142 C210.673,356.119 213.876,344.152 220.319,333.96 L209.066,311.498 C201.398,311.183 194.006,309.107 186.908,306.318 L178.235,313.956 C176.191,338.213 155.394,355.877 131.593,356.479 C105.692,356.479 84.686,335.475 84.686,309.574 C85.016,284.389 103.442,265.336 128.195,262.826 L110.794,260.482 z M120.536,192.785 C120.718,193.162 120.582,192.871 120.89,193.683 L121.496,193.765 L120.536,192.785 z"/>
-      </g>
-      <g>
-        <g fill={props.fill || '#ea9122'}>
-          <path d="M304.391,98.412 L386.004,179.165 C390.211,183.328 392.579,189.001 392.579,194.92 L392.578,377.836 C392.579,390.077 382.656,400 370.415,400 L309.992,400 C318.006,391.135 322.01,380.003 322.311,368.142 C322.311,341.983 301.094,320.767 274.938,320.767 C272.714,320.767 270.5,320.921 268.297,321.233 L247.801,280.319 C258.827,270.437 265.228,256.323 265.228,241.296 C265.228,223.395 256.128,206.933 241.368,197.342 L260.087,135.66 C281.135,134.686 300.219,119.34 304.391,98.412 z M309.477,128.766 C300.404,139.605 288.081,148.591 274.042,151.66 L262.086,191.055 C275.957,204.909 282.613,221.841 283.228,241.296 C283.258,256.463 278.225,270.931 269.341,283.128 L279.27,302.95 C314.279,305.901 339.426,333.165 340.311,368.142 C340.311,368.294 340.309,368.446 340.305,368.598 C340.298,373.126 339.528,377.575 338.66,382 L370.415,382 C372.715,382 374.579,380.136 374.578,377.836 L374.579,194.919 C374.579,193.808 374.134,192.742 373.343,191.96 L309.477,128.766 z"/>
-          <path d="M218.847,113.177 C222.017,118.311 226.142,122.802 230.99,126.395 L212.014,188.919 C204.747,189.172 197.654,190.918 191.137,194.057 L145.569,147.529 C148.378,142.185 150.127,136.37 150.738,130.356 L218.847,113.177 z M209.869,134.005 L167.888,144.594 L195.977,173.274 L198.111,172.748 L209.869,134.005 z"/>
-          <path d="M196.805,-0 L198.918,0.05 C204.164,0.561 208.531,2.9 212.397,6.359 L248.577,42.159 C229.253,46.857 215.302,62.448 212.229,82 L143.141,99.424 C135.178,87.343 122.504,80.259 108.336,78.475 L181.219,6.359 C184.893,2.811 189.589,0.539 194.697,0.05 L196.805,-0 z M196.806,18.005 L196.137,18.021 C195.192,18.113 194.433,18.714 193.736,19.296 L140.687,71.787 C147.148,76.679 144.865,74.688 149.61,79.229 L197.474,67.158 C201.581,55.517 208.76,45.333 218.317,37.539 L200.2,19.613 C199.406,18.9 198.506,18.238 197.442,18.02 L196.806,18.005 z"/>
-        </g>
-      </g>
-      <g fill={props.fill || '#ea9122'}>
-        <path d="M434.037,-0 Q458.795,-0 472.461,13.045 Q486.126,26.089 487.635,42.683 L470.42,42.683 Q467.491,30.082 458.751,22.717 Q450.01,15.352 434.214,15.352 Q414.958,15.352 403.112,28.884 Q391.265,42.417 391.265,70.37 Q391.265,93.264 401.958,107.507 Q412.651,121.749 433.859,121.749 Q453.382,121.749 463.587,106.752 Q469,98.855 471.662,85.988 L488.877,85.988 Q486.57,106.575 473.614,120.507 Q458.085,137.278 431.73,137.278 Q409.013,137.278 393.572,123.524 Q373.251,105.333 373.251,67.352 Q373.251,38.512 388.514,20.055 Q405.019,-0 434.037,-0 z"/>
-        <path d="M567.588,38.867 L585.247,38.867 Q581.875,48.008 570.25,80.575 Q561.554,105.066 555.697,120.507 Q541.854,156.89 536.175,164.876 Q530.496,172.863 516.652,172.863 Q513.28,172.863 511.461,172.596 Q509.642,172.33 506.98,171.62 L506.98,157.067 Q511.151,158.221 513.014,158.487 Q514.878,158.753 516.297,158.753 Q520.734,158.753 522.82,157.289 Q524.905,155.825 526.325,153.695 Q526.769,152.985 529.52,146.418 Q532.27,139.852 533.513,136.657 L498.372,38.867 L516.475,38.867 L541.943,116.247 z"/>
-        <path d="M602.285,12.335 L618.435,12.335 L618.435,38.867 L633.61,38.867 L633.61,51.912 L618.435,51.912 L618.435,113.94 Q618.435,118.91 621.807,120.596 Q623.671,121.572 628.019,121.572 Q629.173,121.572 630.504,121.527 Q631.835,121.483 633.61,121.305 L633.61,133.906 Q630.859,134.705 627.886,135.06 Q624.913,135.415 621.452,135.415 Q610.271,135.415 606.278,129.691 Q602.285,123.968 602.285,114.828 L602.285,51.912 L589.418,51.912 L589.418,38.867 L602.285,38.867 z"/>
-        <path d="M687.296,123.613 Q703.181,123.613 709.082,111.589 Q714.983,99.565 714.983,84.834 Q714.983,71.523 710.723,63.182 Q703.979,50.049 687.474,50.049 Q672.832,50.049 666.177,61.23 Q659.521,72.411 659.521,88.206 Q659.521,103.38 666.177,113.496 Q672.832,123.613 687.296,123.613 z M687.918,36.117 Q706.286,36.117 718.976,48.362 Q731.666,60.608 731.666,84.39 Q731.666,107.374 720.485,122.37 Q709.304,137.367 685.788,137.367 Q666.177,137.367 654.641,124.101 Q643.105,110.834 643.105,88.472 Q643.105,64.513 655.262,50.315 Q667.419,36.117 687.918,36.117 z"/>
-        <path d="M760.151,104.09 Q760.861,112.077 764.144,116.336 Q770.178,124.056 785.086,124.056 Q793.96,124.056 800.704,120.196 Q807.448,116.336 807.448,108.261 Q807.448,102.138 802.035,98.943 Q798.574,96.991 788.37,94.418 L775.68,91.223 Q763.523,88.206 757.755,84.479 Q747.461,78.001 747.461,66.554 Q747.461,53.066 757.178,44.724 Q766.895,36.383 783.311,36.383 Q804.786,36.383 814.281,48.984 Q820.227,56.97 820.049,66.199 L804.964,66.199 Q804.52,60.786 801.148,56.349 Q795.646,50.048 782.069,50.048 Q773.018,50.048 768.359,53.509 Q763.7,56.97 763.7,62.649 Q763.7,68.861 769.823,72.588 Q773.373,74.807 780.294,76.493 L790.854,79.066 Q808.069,83.237 813.926,87.141 Q823.244,93.264 823.244,106.397 Q823.244,119.087 813.616,128.316 Q803.987,137.545 784.288,137.545 Q763.079,137.545 754.25,127.916 Q745.42,118.288 744.799,104.09 z"/>
-        <path d="M878.173,36.117 Q894.234,36.117 904.306,43.926 Q914.378,51.735 916.419,70.813 L900.89,70.813 Q899.47,62.028 894.412,56.216 Q889.354,50.403 878.173,50.403 Q862.91,50.403 856.343,65.312 Q852.084,74.984 852.084,89.182 Q852.084,103.469 858.118,113.23 Q864.152,122.991 877.108,122.991 Q887.047,122.991 892.859,116.913 Q898.671,110.834 900.89,100.274 L916.419,100.274 Q913.757,119.176 903.108,127.916 Q892.46,136.657 875.866,136.657 Q857.231,136.657 846.138,123.036 Q835.046,109.415 835.046,89.005 Q835.046,63.98 847.203,50.048 Q859.36,36.117 878.173,36.117 z"/>
-        <path d="M944.638,108.616 Q944.638,115.537 949.696,119.531 Q954.754,123.524 961.676,123.524 Q970.106,123.524 978.004,119.619 Q991.314,113.142 991.314,98.411 L991.314,85.544 Q988.386,87.407 983.772,88.65 Q979.157,89.892 974.72,90.424 L965.048,91.667 Q956.351,92.82 952.003,95.305 Q944.638,99.476 944.638,108.616 z M983.328,76.315 Q988.83,75.605 990.693,71.701 Q991.758,69.571 991.758,65.578 Q991.758,57.414 985.946,53.731 Q980.133,50.048 969.307,50.048 Q956.795,50.048 951.56,56.793 Q948.631,60.52 947.744,67.885 L932.836,67.885 Q933.279,50.315 944.239,43.437 Q955.198,36.56 969.662,36.56 Q986.434,36.56 996.905,42.949 Q1007.287,49.339 1007.287,62.827 L1007.287,117.578 Q1007.287,120.063 1008.308,121.572 Q1009.328,123.08 1012.612,123.08 Q1013.676,123.08 1015.008,122.947 Q1016.339,122.814 1017.847,122.548 L1017.847,134.35 Q1014.12,135.415 1012.168,135.681 Q1010.216,135.947 1006.844,135.947 Q998.591,135.947 994.864,130.091 Q992.912,126.985 992.113,121.305 Q987.232,127.695 978.092,132.398 Q968.952,137.101 957.949,137.101 Q944.727,137.101 936.341,129.07 Q927.955,121.039 927.955,108.971 Q927.955,95.749 936.208,88.472 Q944.46,81.196 957.86,79.51 z"/>
-        <path d="M1073.575,123.169 Q1084.756,123.169 1092.165,113.807 Q1099.575,104.445 1099.575,85.81 Q1099.575,74.452 1096.292,66.288 Q1090.08,50.581 1073.575,50.581 Q1056.981,50.581 1050.858,67.175 Q1047.574,76.049 1047.574,89.715 Q1047.574,100.718 1050.858,108.438 Q1057.069,123.169 1073.575,123.169 z M1032.223,39.311 L1047.752,39.311 L1047.752,51.912 Q1052.544,45.434 1058.223,41.885 Q1066.298,36.56 1077.213,36.56 Q1093.363,36.56 1104.633,48.939 Q1115.903,61.318 1115.903,84.302 Q1115.903,115.36 1099.664,128.671 Q1089.37,137.101 1075.704,137.101 Q1064.967,137.101 1057.691,132.398 Q1053.431,129.736 1048.196,123.258 L1048.196,171.798 L1032.223,171.798 z"/>
-        <path d="M1174.115,36.738 Q1184.232,36.738 1193.727,41.485 Q1203.222,46.233 1208.191,53.776 Q1212.983,60.963 1214.58,70.547 Q1216,77.114 1216,91.489 L1146.34,91.489 Q1146.784,105.954 1153.173,114.694 Q1159.562,123.435 1172.962,123.435 Q1185.474,123.435 1192.928,115.183 Q1197.188,110.391 1198.962,104.09 L1214.669,104.09 Q1214.048,109.326 1210.543,115.759 Q1207.037,122.193 1202.689,126.275 Q1195.413,133.374 1184.675,135.859 Q1178.907,137.278 1171.631,137.278 Q1153.883,137.278 1141.548,124.367 Q1129.214,111.456 1129.214,88.206 Q1129.214,65.312 1141.637,51.025 Q1154.061,36.738 1174.115,36.738 z M1199.583,78.8 Q1198.607,68.417 1195.058,62.206 Q1188.491,50.67 1173.139,50.67 Q1162.136,50.67 1154.682,58.612 Q1147.228,66.554 1146.784,78.8 z"/>
-      </g>
-      <g fill={props.fill || '#000000'}>
-        <path d="M487.304,395.987 L487.304,189.591 L506.797,189.591 L506.797,278.169 L628.914,278.169 L628.914,189.591 L648.407,189.591 L648.407,395.987 L628.914,395.987 L628.914,294.796 L506.797,294.796 L506.797,395.987 z"/>
-        <path d="M751.032,400 Q736.699,399.713 724.229,394.84 Q711.759,389.967 702.443,380.22 Q693.126,370.474 687.68,355.854 Q682.233,341.235 682.233,321.742 Q682.233,305.402 686.533,291.212 Q690.833,277.023 699.433,266.416 Q708.033,255.81 720.932,249.646 Q733.832,243.483 751.032,243.483 Q768.518,243.483 781.418,249.646 Q794.318,255.81 802.917,266.273 Q811.517,276.736 815.674,291.069 Q819.83,305.402 819.83,321.742 Q819.83,339.515 814.957,353.848 Q810.084,368.181 801.054,378.5 Q792.024,388.82 779.268,394.41 Q766.511,400 751.032,400 z M801.484,321.742 Q801.484,308.555 798.617,297.089 Q795.751,285.622 789.731,277.166 Q783.711,268.709 774.251,263.836 Q764.792,258.963 751.032,258.963 Q737.559,258.963 727.956,264.266 Q718.352,269.569 712.333,278.313 Q706.313,287.056 703.446,298.379 Q700.579,309.702 700.579,321.742 Q700.579,335.501 703.876,347.111 Q707.173,358.721 713.479,367.034 Q719.786,375.347 729.246,379.934 Q738.705,384.52 751.032,384.52 Q764.505,384.52 774.108,379.217 Q783.711,373.914 789.731,365.027 Q795.751,356.141 798.617,344.961 Q801.484,333.781 801.484,321.742 z"/>
-        <path d="M866.269,395.987 L849.07,395.987 L849.07,247.497 L866.269,247.497 L866.269,269.856 L866.843,269.856 Q873.436,256.956 885.762,250.22 Q898.089,243.483 912.708,243.483 Q923.028,243.483 930.195,245.777 Q937.361,248.07 942.378,251.94 Q947.394,255.81 950.691,260.97 Q953.988,266.13 956.281,272.149 Q963.734,257.53 976.204,250.507 Q988.673,243.483 1004.727,243.483 Q1019.919,243.483 1028.949,248.5 Q1037.979,253.516 1042.852,261.686 Q1047.726,269.856 1049.302,280.463 Q1050.879,291.069 1050.879,302.249 L1050.879,395.987 L1033.679,395.987 L1033.679,293.362 Q1033.679,277.309 1026.943,268.136 Q1020.206,258.963 1001.86,258.963 Q992.4,258.963 983.657,263.12 Q974.914,267.276 968.321,275.876 Q965.454,279.603 963.591,284.619 Q961.727,289.636 960.581,295.082 Q959.434,300.529 959.004,305.975 Q958.574,311.422 958.574,316.008 L958.574,395.987 L941.374,395.987 L941.374,293.362 Q941.374,277.309 934.638,268.136 Q927.901,258.963 909.555,258.963 Q900.095,258.963 891.352,263.12 Q882.609,267.276 876.016,275.876 Q873.149,279.603 871.286,284.619 Q869.423,289.636 868.276,295.082 Q867.129,300.529 866.699,305.975 Q866.269,311.422 866.269,316.008 z"/>
-        <path d="M1197.649,311.422 Q1197.649,300.529 1194.639,290.926 Q1191.629,281.322 1185.753,274.299 Q1179.876,267.276 1170.99,263.12 Q1162.103,258.963 1150.637,258.963 Q1139.17,258.963 1130.427,263.263 Q1121.684,267.563 1115.521,274.729 Q1109.358,281.896 1105.631,291.356 Q1101.904,300.815 1100.758,311.422 z M1100.758,325.755 Q1100.758,337.795 1103.911,348.544 Q1107.064,359.294 1113.228,367.177 Q1119.391,375.061 1128.851,379.79 Q1138.31,384.52 1150.637,384.52 Q1167.55,384.52 1179.589,375.634 Q1191.629,366.747 1195.642,350.408 L1214.849,350.408 Q1211.695,361.014 1206.536,370.044 Q1201.376,379.074 1193.636,385.81 Q1185.896,392.547 1175.29,396.273 Q1164.683,400 1150.637,400 Q1132.864,400 1119.964,393.98 Q1107.064,387.96 1098.751,377.497 Q1090.438,367.034 1086.425,352.844 Q1082.411,338.655 1082.411,322.315 Q1082.411,305.975 1087.141,291.642 Q1091.871,277.309 1100.615,266.559 Q1109.358,255.81 1121.971,249.646 Q1134.584,243.483 1150.637,243.483 Q1181.309,243.483 1198.796,264.553 Q1216.282,285.622 1215.995,325.755 z"/>
-      </g>
-    </svg>
-  )
-}
-
-export function CytoscapeLogo(props) {
-  return (
-    <svg viewBox="0 0 40 40" aria-hidden="true" {...props}>
-      <path d="M18.13,9.547 C18.065,10.19 17.878,10.811 17.578,11.381 L22.444,16.35 C23.141,16.015 23.898,15.828 24.674,15.801 L26.701,9.124 C26.183,8.741 25.742,8.261 25.404,7.713 L18.13,9.547 z M17.319,6.244 L24.697,4.383 C25.028,1.903 27.154,0 29.711,0 C32.505,0 34.77,2.266 34.77,5.059 C34.77,7.783 32.616,10.005 29.919,10.114 L27.92,16.701 C29.496,17.725 30.468,19.483 30.468,21.395 C30.468,23 29.784,24.507 28.607,25.562 L30.795,29.931 C31.03,29.898 31.267,29.882 31.505,29.882 C34.298,29.882 36.563,32.147 36.563,34.941 C36.563,37.734 34.298,40 31.505,40 C28.711,40 26.445,37.734 26.445,34.941 C26.445,33.647 26.936,32.418 27.809,31.486 L25.538,26.952 C25.316,26.979 25.093,26.992 24.871,26.992 C23.737,26.992 22.644,26.652 21.718,26.019 L19.146,28.284 C19.163,28.417 19.172,28.552 19.172,28.686 C19.172,30.391 17.79,31.773 16.086,31.773 C14.381,31.773 12.999,30.391 12.999,28.686 C12.999,26.982 14.381,25.599 16.086,25.599 C16.38,25.599 16.671,25.641 16.951,25.723 L19.631,23.364 C19.494,23.001 19.396,22.625 19.338,22.241 L12.951,21.381 C12.078,23.005 10.374,24.043 8.496,24.043 C5.702,24.043 3.437,21.777 3.437,18.984 C3.437,16.191 5.702,13.925 8.496,13.925 C10.953,13.925 13.029,15.688 13.468,18.051 L19.854,18.912 C19.896,18.826 19.941,18.741 19.988,18.658 L15.109,13.675 C14.476,13.95 13.793,14.092 13.097,14.092 C10.304,14.092 8.038,11.826 8.038,9.032 C8.038,6.239 10.304,3.974 13.097,3.974 C14.812,3.973 16.39,4.838 17.319,6.244 z" fill={props.fill || '#ea9122'}/>
-    </svg>
-  )
-}
-
-export function NDExLogo(props) {
-  return (
-    <svg viewBox="0 0 32 32" aria-hidden="true" {...props}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M30.385,17.995 C30.118,18.255 29.839,18.493 29.547,18.71 C29.255,18.927 28.967,19.125 28.682,19.306 C27.783,19.785 26.918,20.344 25.987,20.745 C25.687,20.875 25.381,21.009 25.07,21.145 C24.76,21.282 24.451,21.41 24.144,21.529 C23.837,21.648 23.528,21.776 23.218,21.914 C22.591,22.13 21.969,22.342 21.351,22.551 C20.735,22.76 20.106,22.962 19.467,23.156 C19.151,23.235 18.84,23.325 18.537,23.426 C17.618,23.731 16.623,23.998 15.679,24.193 C15.402,24.243 15.142,24.18 14.898,24.002 C14.656,23.824 14.513,23.58 14.471,23.269 C14.429,22.958 14.48,22.663 14.624,22.385 C14.853,21.945 15.36,21.843 15.78,21.784 C16.537,21.677 17.33,21.42 18.08,21.245 C18.386,21.173 18.697,21.097 19.014,21.018 C19.634,20.838 20.249,20.662 20.859,20.489 C21.47,20.316 22.079,20.129 22.686,19.928 C23.293,19.726 23.897,19.519 24.498,19.306 C25.664,18.892 26.853,18.382 27.943,17.754 C28.777,17.269 30.03,16.542 30.41,15.457 C31.436,11.473 18.639,11.752 19.443,11.505 C22.638,11.264 27.067,10.721 30.149,12.43 C30.239,12.487 30.32,12.543 30.393,12.598 C30.467,12.652 30.549,12.709 30.639,12.767 C32.632,14.206 31.833,16.606 30.385,17.995 z"
-        fill="#00a1de"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M1.615,14.515 C0.167,15.904 -0.632,18.304 1.362,19.744 C1.451,19.801 1.533,19.858 1.607,19.912 C1.68,19.967 1.762,20.023 1.851,20.08 C4.933,21.789 9.362,21.246 12.557,21.005 C13.361,20.758 0.564,21.038 1.59,17.053 C1.97,15.968 3.224,15.241 4.057,14.757 C5.147,14.128 6.335,13.618 7.501,13.204 C8.103,12.991 8.707,12.785 9.314,12.583 C9.921,12.381 10.531,12.194 11.141,12.022 C11.751,11.848 12.366,11.672 12.986,11.492 C13.302,11.413 13.614,11.337 13.921,11.265 C14.67,11.09 15.463,10.833 16.22,10.726 C16.64,10.667 17.147,10.565 17.376,10.125 C17.52,9.847 17.572,9.552 17.529,9.241 C17.487,8.93 17.344,8.686 17.102,8.508 C16.858,8.331 16.598,8.268 16.321,8.318 C15.377,8.512 14.382,8.779 13.463,9.084 C13.159,9.185 12.85,9.275 12.533,9.354 C11.894,9.548 11.265,9.751 10.649,9.959 C10.031,10.168 9.409,10.381 8.783,10.597 C8.472,10.734 8.163,10.862 7.856,10.981 C7.549,11.1 7.24,11.228 6.93,11.365 C6.619,11.502 6.314,11.636 6.014,11.765 C5.082,12.166 4.218,12.725 3.318,13.205 C3.033,13.385 2.745,13.583 2.453,13.8 C2.161,14.017 1.882,14.255 1.615,14.515 z"
-        fill="#00a1de"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M19.695,7 C20.529,7.18 21.005,7.781 21.111,8.6 C21.194,9.586 20.548,10.467 19.67,10.569 C19.427,10.598 19.191,10.563 18.976,10.477 L16.58,15.12 C16.672,15.22 16.746,15.342 16.793,15.48 L18.63,15.121 C18.648,14.674 18.954,14.296 19.364,14.249 C19.809,14.197 20.205,14.56 20.246,15.06 C20.288,15.56 19.961,16.008 19.515,16.059 C19.135,16.104 18.792,15.846 18.673,15.459 L16.853,15.815 C16.863,16.16 16.717,16.477 16.483,16.673 L17.135,18.356 C17.579,18.235 18.39,18.726 18.446,19.4 C18.501,20.06 18.069,20.651 17.48,20.72 C16.891,20.788 16.361,20.206 16.314,19.648 C16.267,19.09 16.469,18.77 16.786,18.529 L16.136,16.852 C16.102,16.861 16.066,16.867 16.031,16.871 C15.926,16.884 15.823,16.875 15.727,16.849 L13.348,21.702 C13.71,21.994 13.962,22.455 14.007,22.994 C14.09,23.993 13.436,24.887 12.546,24.99 C11.656,25.094 10.866,24.368 10.782,23.369 C10.699,22.37 11.353,21.477 12.243,21.373 C12.465,21.347 12.682,21.373 12.882,21.442 L15.277,16.555 C15.183,16.441 15.112,16.302 15.072,16.147 L13.649,16.426 C13.648,16.795 13.396,17.114 13.062,17.153 C12.706,17.195 12.39,16.904 12.357,16.505 C12.323,16.105 12.585,15.748 12.941,15.706 C13.248,15.67 13.525,15.881 13.617,16.196 L15.038,15.918 L15.038,15.919 C15.03,15.73 15.068,15.549 15.142,15.391 L13.669,14.221 C13.539,14.366 13.366,14.463 13.17,14.486 C12.717,14.539 12.316,14.17 12.274,13.662 C12.231,13.155 12.564,12.701 13.016,12.648 C13.468,12.595 13.87,12.964 13.912,13.472 C13.927,13.646 13.897,13.814 13.834,13.961 L15.311,15.135 C15.412,15.022 15.537,14.935 15.678,14.885 L15.452,13.34 C14.976,13.361 14.566,12.966 14.521,12.426 C14.474,11.866 14.841,11.365 15.34,11.306 C15.84,11.248 16.283,11.655 16.329,12.216 C16.368,12.681 16.121,13.106 15.75,13.27 L15.981,14.837 C16.027,14.838 16.073,14.844 16.117,14.853 L18.521,10.195 C18.195,9.905 17.972,9.472 17.93,8.97 C17.847,7.985 18.493,7.103 19.371,7.001 L19.695,7 z M12.521,21.889 L12.287,21.89 C11.65,21.964 11.183,22.602 11.243,23.315 C11.303,24.029 11.867,24.548 12.503,24.474 C13.139,24.4 13.606,23.761 13.546,23.048 C13.441,22.448 13.142,22.022 12.521,21.889 z M17.463,18.669 L17.308,18.669 C16.888,18.718 16.578,19.141 16.618,19.613 C16.658,20.084 17.031,20.427 17.452,20.378 C17.872,20.329 18.181,19.907 18.142,19.435 C18.072,19.039 17.874,18.757 17.463,18.669 z M13.052,15.912 L12.958,15.913 C12.704,15.942 12.517,16.198 12.541,16.483 C12.565,16.769 12.791,16.976 13.045,16.947 C13.299,16.917 13.486,16.662 13.463,16.376 C13.42,16.136 13.301,15.966 13.052,15.912 z M16.016,15.132 L15.885,15.132 C15.528,15.174 15.266,15.532 15.3,15.932 C15.333,16.332 15.65,16.623 16.006,16.582 C16.363,16.54 16.625,16.182 16.592,15.782 C16.533,15.445 16.365,15.206 16.016,15.132 z M19.502,14.507 L19.385,14.508 C19.067,14.545 18.833,14.864 18.863,15.221 C18.893,15.578 19.175,15.838 19.494,15.801 C19.812,15.764 20.046,15.444 20.016,15.087 C19.963,14.787 19.813,14.574 19.502,14.507 z M13.157,12.91 L13.038,12.91 C12.715,12.948 12.477,13.273 12.508,13.635 C12.538,13.998 12.825,14.261 13.148,14.224 C13.471,14.186 13.709,13.862 13.678,13.499 C13.625,13.194 13.472,12.978 13.157,12.91 z M15.496,11.596 L15.364,11.596 C15.008,11.637 14.745,11.996 14.779,12.396 C14.812,12.796 15.129,13.087 15.486,13.046 C15.842,13.004 16.105,12.646 16.071,12.246 C16.012,11.909 15.844,11.67 15.496,11.596 z M19.645,7.51 L19.414,7.511 C18.787,7.584 18.326,8.214 18.385,8.917 C18.443,9.621 19,10.133 19.627,10.06 C20.255,9.987 20.716,9.357 20.657,8.653 C20.553,8.062 20.257,7.641 19.645,7.51 z"
-        fill="#00a1de"
-      />
-    </svg>
-  )
-}
-
-export function GeneManiaLogo(props) {
-  return (
-    <svg viewBox="0 0 32 32" aria-hidden="true" {...props}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M14.774,1 C21.322,1.016 26.534,4.704 29.056,10.682 C31.977,18.726 27.492,27.819 19.369,30.725 C18.625,30.991 17.807,30.604 17.541,29.86 C17.275,29.116 17.662,28.298 18.406,28.032 C25.059,25.652 28.74,18.189 26.368,11.659 C24.185,5.652 17.426,2.347 11.546,4.522 C6.184,6.505 3.256,12.559 5.233,17.79 C7.014,22.506 12.364,25.06 16.945,23.283 C21.017,21.702 23.198,17.058 21.621,13.127 C20.247,9.698 16.308,7.89 13.028,9.263 C10.239,10.43 8.801,13.661 9.965,16.289 C10.919,18.438 13.443,19.513 15.416,18.564 C16.932,17.833 17.649,16.023 16.93,14.705 C16.352,13.831 15.944,13.686 14.953,13.767 C15.882,14.267 15.579,13.923 15.983,14.678 C16.361,15.942 15.53,16.921 14.398,17.061 C13.74,17.143 12.983,16.955 12.486,16.526 C10.847,15.107 11.454,12.646 13.095,11.525 C15.263,10.047 18.202,11.064 19.44,13.335 C20.988,16.17 19.576,19.734 16.657,21.141 C13.166,22.82 8.934,21.017 7.351,17.448 C5.514,13.305 7.702,8.391 11.924,6.625 C16.718,4.618 22.323,7.191 24.276,12.063 C26.459,17.507 23.502,23.805 17.979,25.949 C11.885,28.313 4.889,24.973 2.557,18.801 C0.008,12.057 3.73,4.364 10.554,1.84 C12.858,1.125 11.468,1.476 14.774,1 z"
-        fill="#000000"
-      />
-    </svg>
-  )
-}
-
-export function EnrichmentMapLogo(props) {
-  return (
-    <svg viewBox="0 0 32 32" aria-hidden="true" {...props}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M-0,7.607 L32,7.607 L32,24.393 L-0,24.393 L-0,7.607 z"
-        fill="#ffffff"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M16,8.393 L16,12.385 L9.138,12.385 L9.138,14.073 L16,14.073 L16,17.859 L9.138,17.859 L9.138,19.615 L16,19.615 L16,23.607 L0.787,23.607 L0.787,8.393 L16,8.393 z"
-        fill="#1f78b4"/>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M31.213,23.607 L27.222,23.607 L27.222,16.745 L25.534,16.745 L25.534,23.607 L21.748,23.607 L21.748,16.745 L19.991,16.745 L19.991,23.607 L16,23.607 L16,8.393 L31.213,8.393 L31.213,23.607 z"
-        fill="#33a02c"
-      />
-    </svg>
-  )
-}
-
-export function WikiPathwaysLogo(props) {
-  return (
-    <svg viewBox="0 0 32 32" aria-hidden="true" {...props}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M6.968,11.625 L6.968,14.698 L5.955,14.698 L7.365,17.831 L8.741,14.698 L7.728,14.698 L7.728,11.625 L6.968,11.625 z M23.527,17.311 C24.038,17.289 24.444,16.869 24.447,16.357 C24.447,15.83 24.02,15.403 23.493,15.403 C22.966,15.403 22.539,15.83 22.539,16.357 C22.539,16.884 22.966,17.311 23.493,17.311 z M21.433,11.105 L25.579,11.105 L25.579,9.948 L21.399,9.948 z M13.8,6.165 C13.8,5.911 13.698,5.667 13.518,5.488 C13.337,5.309 13.092,5.21 12.837,5.211 C12.447,5.211 12.096,5.447 11.947,5.807 C11.799,6.169 11.883,6.583 12.16,6.858 C12.437,7.132 12.852,7.212 13.212,7.061 C13.571,6.909 13.803,6.555 13.8,6.165 z M7.729,27.047 C9.148,28.098 10.752,28.872 12.457,29.326 L12.457,26.371 C11.525,26.371 10.769,25.615 10.769,24.682 C10.769,23.75 11.525,22.994 12.457,22.994 L12.457,19.827 L10.194,19.827 L10.194,20.41 C10.194,20.62 10.024,20.79 9.814,20.79 L7.729,20.79 z M5.255,20.03 L9.476,20.03 L9.476,18.865 L5.255,18.865 z M13.8,24.682 C13.795,24.155 13.366,23.729 12.837,23.729 C12.447,23.729 12.096,23.964 11.947,24.324 C11.799,24.686 11.883,25.101 12.16,25.375 C12.437,25.649 12.852,25.729 13.212,25.578 C13.571,25.427 13.803,25.073 13.8,24.682 z M26.601,14.964 L29.742,16.374 L26.601,17.75 L26.601,16.737 L25.224,16.737 C25.224,17.67 24.468,18.426 23.535,18.426 C22.603,18.426 21.847,17.67 21.847,16.737 L19.448,16.737 L19.448,15.977 L21.847,15.977 C21.987,15.331 22.493,14.826 23.139,14.685 L23.139,11.873 L21.07,11.873 C20.859,11.869 20.69,11.697 20.69,11.485 L20.69,10.911 L13.8,10.911 L13.8,10.142 L20.673,10.142 L20.673,9.56 C20.673,9.35 20.844,9.18 21.053,9.18 L25.993,9.18 C26.205,9.18 26.377,9.349 26.381,9.56 L26.381,10.142 L28.83,10.142 C27.015,6 23.298,2.998 18.866,2.096 L18.866,9.56 L18.106,9.56 L18.106,1.96 C17.416,1.859 16.719,1.808 16.021,1.808 C15.082,1.807 14.146,1.9 13.226,2.087 L13.226,4.46 C14.158,4.46 14.914,5.216 14.914,6.148 C14.914,7.082 14.158,7.837 13.226,7.837 L13.226,15.977 L17.515,15.977 L17.515,16.737 L12.837,16.737 C12.628,16.737 12.457,16.567 12.457,16.357 L12.457,7.871 C11.525,7.871 10.769,7.115 10.769,6.182 C10.769,5.25 11.525,4.494 12.457,4.494 L12.457,2.264 C10.751,2.717 9.146,3.49 7.729,4.544 L7.729,9.56 L6.96,9.56 L6.96,5.152 C5.359,6.517 4.084,8.221 3.228,10.142 L11.875,10.142 L11.875,10.911 L2.958,10.911 C0.897,16.413 2.496,22.618 6.96,26.439 L6.96,20.79 L4.875,20.79 C4.664,20.79 4.491,20.621 4.486,20.41 L4.486,18.485 C4.491,18.274 4.664,18.105 4.875,18.105 L9.814,18.105 C10.024,18.105 10.194,18.275 10.194,18.485 L10.194,19.067 L12.837,19.067 C13.049,19.067 13.221,19.236 13.226,19.448 L13.226,22.994 C13.863,23.146 14.361,23.641 14.517,24.277 L18.072,24.277 L18.072,11.485 L18.832,11.485 L18.832,24.699 C18.832,24.91 18.663,25.08 18.452,25.08 L14.484,25.08 C14.333,25.72 13.833,26.22 13.192,26.371 L13.192,29.538 C14.112,29.725 15.049,29.819 15.987,29.816 C16.468,29.816 16.873,29.749 17.312,29.749 L17.118,28.71 L20.496,29.462 L17.701,31.438 L17.507,30.517 C17.014,30.568 16.517,30.593 16.021,30.593 C8.589,30.583 2.323,25.049 1.396,17.675 C0.469,10.301 5.169,3.388 12.368,1.538 C19.566,-0.311 27.017,3.48 29.758,10.387 C29.801,10.506 29.785,10.637 29.716,10.742 C29.647,10.847 29.529,10.91 29.404,10.911 L26.381,10.911 L26.381,11.485 C26.377,11.698 26.205,11.869 25.993,11.873 L23.907,11.873 L23.907,14.685 C24.554,14.826 25.059,15.331 25.199,15.977 L26.601,15.977 z"
-        fill="#000000"
-      />
-    </svg>
-  )
-}
-
-export function CytoscapeWebLogo(props) {
-  return (
-    <svg viewBox="0 0 32 32" aria-hidden="true" {...props}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M23.013,0 C25.89,-0 28.224,2.333 28.224,5.21 C28.209,7.495 26.637,9.691 24.372,10.237 L23.507,13.09 C24.571,14.166 25.069,15.51 25.116,17.007 C25.109,18.167 24.758,19.31 24.073,20.25 L24.752,21.603 C27.496,21.851 29.45,24.068 29.518,26.79 C29.518,29.667 27.185,32 24.308,32 C21.431,32 19.097,29.667 19.097,26.79 C19.112,25.872 19.327,24.928 19.822,24.144 L19.04,22.582 C18.502,22.547 17.974,22.413 17.474,22.217 L16.931,22.694 C16.738,24.641 15.084,26.011 13.173,26.059 C11.083,26.059 9.387,24.363 9.387,22.273 C9.405,21.353 9.718,20.379 10.382,19.715 C9.579,20.241 8.639,20.444 7.692,20.477 C4.815,20.477 2.482,18.143 2.482,15.266 C2.508,12.872 4.021,10.956 6.283,10.253 C5.954,9.572 5.824,8.83 5.805,8.08 C5.805,5.203 8.138,2.869 11.015,2.869 C12.418,2.88 13.586,3.386 14.613,4.322 L18.115,3.439 C18.804,1.353 20.878,0.049 23.013,0 z M23.013,1.557 C21.167,1.557 19.631,2.931 19.392,4.723 L14.064,6.067 C13.393,5.051 12.253,4.427 11.015,4.427 C8.998,4.427 7.362,6.063 7.362,8.08 C7.362,10.097 8.998,11.733 11.015,11.733 C11.517,11.733 12.011,11.631 12.468,11.433 L15.991,15.031 C15.958,15.091 15.925,15.152 15.895,15.214 L11.283,14.592 C10.966,12.886 9.467,11.613 7.692,11.613 C5.675,11.613 4.039,13.249 4.039,15.266 C4.039,17.283 5.675,18.92 7.692,18.92 C9.048,18.92 10.279,18.17 10.91,16.997 L15.522,17.619 C15.564,17.895 15.635,18.167 15.733,18.429 L13.798,20.133 C13.596,20.074 13.386,20.043 13.173,20.043 C11.942,20.043 10.944,21.042 10.944,22.273 C10.944,23.503 11.942,24.502 13.173,24.502 C14.404,24.502 15.402,23.503 15.402,22.273 C15.402,22.175 15.396,22.079 15.383,21.983 L17.241,20.347 C17.909,20.804 18.699,21.049 19.517,21.049 C19.678,21.049 19.839,21.04 19.999,21.021 L21.639,24.295 C21.009,24.968 20.654,25.855 20.654,26.79 C20.654,28.807 22.291,30.443 24.308,30.443 C26.325,30.443 27.961,28.807 27.961,26.79 C27.961,24.772 26.325,23.136 24.308,23.136 C24.136,23.136 23.965,23.148 23.796,23.172 L22.215,20.017 C23.066,19.255 23.559,18.166 23.559,17.007 C23.559,15.627 22.857,14.357 21.719,13.618 L23.163,8.861 C25.111,8.782 26.666,7.178 26.666,5.21 C26.666,3.193 25.03,1.557 23.013,1.557 z M19.903,7.127 C20.147,7.523 20.465,7.869 20.839,8.146 L19.376,12.968 C18.815,12.987 18.268,13.122 17.765,13.364 L14.251,9.776 C14.468,9.364 14.603,8.916 14.65,8.452 L19.903,7.127 z M18.962,8.97 L16.299,9.642 L18.176,11.558 L18.962,8.97 z M11.635,18.666 L11.36,18.953 C11.535,18.852 11.718,18.764 11.911,18.704 L11.635,18.666 z"
-        fill="#ea9122"
-      />
-    </svg>
-  )
-}
